render error explicitly

+++ app/controllers/sessions_controller.rb
@@ -20,6 +20,7 @@ class SessionsController < ApplicationController
   end
 
   def create
+    raise params[:error] if params[:error]
     if session.delete(:state) == params[:state]
       access_token = current_client.tokenize params[:access_token]
       account = if params[:id_token].present?
